refactor(cli): extract duplicate _CliPoolCore Protocol to shared cli_pool_types module
