+++ src/lyra/core/cli/cli_pool_streaming.py
@@ -8,33 +8,15 @@ from __future__ import annotations
 import asyncio
 import logging
 import time
-from typing import TYPE_CHECKING, Protocol, cast
+from typing import TYPE_CHECKING, cast
 
 from ..agent.agent_config import ModelConfig
+from .cli_pool_types import _CliPoolCore
 from .cli_pool_worker import _ProcessEntry
 from .cli_protocol import StreamingIterator, send_and_read_stream
 
 if TYPE_CHECKING:
     from .cli_protocol import CliProtocolOptions
-
-
-class _CliPoolCore(Protocol):
-    """Protocol declaring cross-mixin dependencies for CliPoolStreamingMixin."""
-
-    async def _idle_reaper(self) -> None: ...
-
-    async def _kill(
-        self, pool_id: str, *, preserve_session: bool = ...
-    ) -> None: ...
-
-    async def _spawn(
-        self,
-        pool_id: str,
-        model_config: ModelConfig,
-        system_prompt: str = ...,
-    ) -> _ProcessEntry | None: ...
-
-    async def reset(self, pool_id: str) -> None: ...
 
 log = logging.getLogger(__name__)
